requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
